ai-service/requirements.txt
# AI Service Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.34
langchain==0.3.7
langchain-openai==0.2.5
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
